+++ RaceManager/DataProcessing/Files/FileManagePolaire.cs
@@ -2,7 +2,7 @@
 {
     public partial class FileManageData
     {
-        public static string CreateFilePolaire(string Name, int ID, string data)
+        public static string CreateFilePolaire(string Name, long ID, string data)
         {
             string pathFile = pathDataPol + Name + "_" + ID + ".pol";
             File.Create(pathFile).Close();
